requirements.txt
langchain==0.3.25
langchain-community==0.3.24
langchain-openai==0.3.16
openai==1.77.0
tiktoken==0.9.0
chromadb==1.0.5
python-dotenv==1.1.0
